sequencing server test - move login user to its own test util js file
