+++ sequencing-server/jestGlobalSetup.js
@@ -3,7 +3,7 @@ import fetch from 'node-fetch';
 import dotenv from 'dotenv';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import { loginTestUser } from "./test/testUtils/testUtils.js";
+import { loginTestUser } from "./test/testUtils/login.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
