+++ sequencing-server/test/testUtils/testUtils.ts
@@ -1,5 +1,4 @@
 import {GraphQLClient} from 'graphql-request';
-import fetch from "node-fetch";
 
 export async function waitMs(ms: number): Promise<void> {
   return new Promise(resolve => setTimeout(resolve, ms));
@@ -13,16 +12,4 @@ export async function getGraphQLClient(): Promise<GraphQLClient> {
       'x-hasura-role': 'aerie_admin',
     },
   });
-}
-
-export async function loginTestUser() {
-  const response = await fetch(`${process.env['MERLIN_GATEWAY_URL']}/auth/login`, {
-    method: 'POST',
-    body: `{"username": "AerieE2ESequencingTests", "password": "password"}`,
-    headers: {'Content-Type': 'application/json'},
-  });
-  if (!response.ok) {
-    throw new Error(`Failed to login: ${response.statusText}`);
-  }
-  return (await response.json() as { token: string }).token;
 }
